Include monster's negative energy resistance in @??.
Currently, only monsters that are wearing artefact armour or wielding
artefact weapons that have rN+ or other forms of equipment with rN+
actually have non-total resistance. Shadow dragons and non-nautral
monsters are rN+++.

Regardless, it's good to show, because sometimes it's not immediately
appararent from the holiness status whether or not a monster will resist
draining.